times: Use 64-bit integer to avoid wasm integer overflow

+++ newlib/libc/machine/cheerp/clientbridge.cpp
@@ -72,7 +72,7 @@ int gettimeofday (timeval* tv, void* tz_)
 	return 0;
 }
 
-[[cheerp::genericjs]] double getPerfNow()
+[[cheerp::genericjs]] uint64_t getPerfNow()
 {
 	return client::performance.now();
 }
